requirements.txt
streamlit==1.28.0
pandas==2.0.3
torch==2.2.1
Pillow==10.1.0
opencv-python-headless==4.9.0.80
numpy==1.26.0
ultralytics==8.3.107